examples: added concept of inheritance, interface and extension
Classes can inherit from each other, override methods, implement
interfaces. Also the programmer can create a class extension by using
`extend ...` block to add functionality to specified class (which
will not be inherited by derived classes) and use `extend ... by ...` to
implement interface for the specified class.

For sure it needs more work and design to make it fluent and easy to
program in objective manner, but it is the first cut of an idea about
this kind of stuff in sloflang.