stage1: change the way tokenizer is called
Until this patch the the Tokenizer instance must have
been created to perform tokenization of input string.
This patch changes that in favour of static method
tokenize() which can be called with input stream.

Also the crude error reporting and returning the result via
TokenizationResult is much more descriptive than the previous
solution to this problem.